requirements.txt
pygame==1.9.6
rpi_ws281x==4.1.0
RPi.GPIO==0.6.5